chore: use command for copy md link
