+++ src/features/NotesContainer/NoteContextMenu/useNoteContextMenu.ts
@@ -1,5 +1,4 @@
 import { useCallback } from 'react';
-import { formatNoteLink } from '@core/features/links';
 import { INote, NoteId } from '@core/features/notes';
 import { TELEMETRY_EVENT_NAME } from '@core/features/telemetry';
 import { ContextMenu } from '@electron/requests/contextMenu';
@@ -12,24 +11,15 @@ import { GLOBAL_COMMANDS } from '@hooks/commands';
 import { useCommand } from '@hooks/commands/useCommand';
 import { ContextMenuCallback } from '@hooks/useContextMenu';
 import { useShowNoteContextMenu } from '@hooks/useShowNoteContextMenu';
-import { useAppSelector } from '@state/redux/hooks';
-import { useVaultSelector, useWorkspaceData } from '@state/redux/profiles/hooks';
-import { selectWorkspace } from '@state/redux/profiles/profiles';
-import { selectDeletionConfig } from '@state/redux/profiles/selectors/vault';
-import { copyTextToClipboard } from '@utils/clipboard';
 
 import { NoteActions } from '.';
+import { useVaultSelector } from '@state/redux/profiles/hooks';
+import { selectDeletionConfig } from '@state/redux/profiles/selectors/vault';
 
 export type ContextMenuOptions = {
 	closeNote: (id: NoteId) => void;
 	updateNotes: () => void;
 };
-
-const mdCharsForEscape = ['\\', '[', ']'];
-const mdCharsForEscapeRegEx = new RegExp(
-	`(${mdCharsForEscape.map((char) => '\\' + char).join('|')})`,
-	'g',
-);
 
 const buildNoteMenu = ({
 	note,
@@ -131,6 +121,8 @@ export const useNoteContextMenu = ({ updateNotes }: ContextMenuOptions) => {
 	const notes = useNotesRegistry();
 	const tagsRegistry = useTagsRegistry();
 
+	const deletionConfig = useVaultSelector(selectDeletionConfig);
+
 	const runCommand = useCommand();
 
 	const noteContextMenuCallback = useCallback<ContextMenuCallback<NoteActions>>(
@@ -175,21 +167,7 @@ export const useNoteContextMenu = ({ updateNotes }: ContextMenuOptions) => {
 				},
 
 				[NoteActions.COPY_MARKDOWN_LINK]: async (id: string) => {
-					const [note] = await notes.getById([id]);
-					if (!note) {
-						console.error(`Can't get data of note #${id}`);
-						return;
-					}
-
-					const { title, text } = note.content;
-					const noteTitle = (title || text.slice(0, 30))
-						.trim()
-						.replace(mdCharsForEscapeRegEx, '\\$1');
-					const markdownLink = `[${noteTitle}](${formatNoteLink(id)})`;
-
-					console.log(`Copy markdown link ${markdownLink}`);
-
-					copyTextToClipboard(markdownLink);
+					runCommand(GLOBAL_COMMANDS.COPY_NOTE_MARKDOWN_LINK, { id });
 				},
 
 				[NoteActions.EXPORT]: async (id: string) => {
